+++ ExilePartyClient/src/app/authorize/components/char-profile/char-wealth/char-wealth.component.ts
@@ -4,12 +4,13 @@ import { Router } from '@angular/router';
 
 import { NetWorthSnapshot } from '../../../../shared/interfaces/income.interface';
 import { Player } from '../../../../shared/interfaces/player.interface';
+import { AccountService } from '../../../../shared/providers/account.service';
 import { AnalyticsService } from '../../../../shared/providers/analytics.service';
 import { ElectronService } from '../../../../shared/providers/electron.service';
+import { IncomeService } from '../../../../shared/providers/income.service';
 import { PartyService } from '../../../../shared/providers/party.service';
 import { SettingsService } from '../../../../shared/providers/settings.service';
 import { NetworthTableComponent } from '../../networth-table/networth-table.component';
-import { AccountService } from '../../../../shared/providers/account.service';
 
 @Component({
   selector: 'app-char-wealth',
@@ -27,6 +28,8 @@ export class CharWealthComponent implements OnInit {
   public graphDimensions = [640, 200];
   public gain = 0;
   public showReset = false;
+  public previousSnapshot = false;
+  public networthValue = 0;
 
   constructor(
     @Inject(FormBuilder) fb: FormBuilder,
@@ -35,6 +38,8 @@ export class CharWealthComponent implements OnInit {
     private partyService: PartyService,
     private analyticsService: AnalyticsService,
     private settingService: SettingsService,
+    private incomeService: IncomeService,
+    private accountService: AccountService
   ) {
     this.form = fb.group({
       searchText: ['']
@@ -46,7 +51,9 @@ export class CharWealthComponent implements OnInit {
         this.showReset = false;
       }
       this.player = res;
+      this.networthValue = this.player.netWorthSnapshots[0].value;
       this.updateGain(res);
+      this.previousSnapshot = false;
     });
   }
 
@@ -63,10 +70,12 @@ export class CharWealthComponent implements OnInit {
   }
 
   resetNetWorth() {
-    if (this.player.account === this.partyService.currentPlayer.account) {
+    const player = this.player;
+    if (player.account === this.partyService.currentPlayer.account) {
       const emptyHistory = this.settingService.deleteNetWorth();
-      this.player.netWorthSnapshots = emptyHistory.history;
-      this.partyService.updatePlayer(this.player);
+      player.netWorthSnapshots = emptyHistory.history;
+      this.incomeService.loadSnapshotsFromSettings();
+      this.accountService.player.next(this.player);
     }
   }
 
@@ -78,6 +87,19 @@ export class CharWealthComponent implements OnInit {
     this.isGraphHidden = false;
   }
 
+  loadPreviousSnapshot(event) {
+    if (this.player.netWorthSnapshots[0] !== undefined) {
+      this.table.loadPreviousSnapshot(event);
+
+      this.networthValue = event.value;
+
+      const lastSnapshotTimestamp = this.player.netWorthSnapshots[0].timestamp;
+      const loadedSnapshotTimestamp = event.name.getTime();
+
+      this.previousSnapshot = loadedSnapshotTimestamp !== lastSnapshotTimestamp;
+    }
+  }
+
   search() {
     this.table.doSearch(this.form.controls.searchText.value);
   }
@@ -87,7 +109,7 @@ export class CharWealthComponent implements OnInit {
     const pastHoursSnapshots = player.netWorthSnapshots
       .filter((snaphot: NetWorthSnapshot) => snaphot.timestamp > oneHourAgo);
 
-    if (pastHoursSnapshots[0] !== undefined) {
+    if (pastHoursSnapshots.length > 1) {
       const lastSnapshot = pastHoursSnapshots[0];
       const firstSnapshot = pastHoursSnapshots[pastHoursSnapshots.length - 1];