+++ ExilePartyClient/src/app/authorize/components/income/income.component.ts
@@ -1,6 +1,7 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { colorSets as ngxChartsColorsets } from '@swimlane/ngx-charts/release/utils/color-sets';
 import * as d3 from 'd3';
+import * as moment from 'moment';
 
 import { ChartSeries, ChartSeriesEntry } from '../../../shared/interfaces/chart.interface';
 import { Player } from '../../../shared/interfaces/player.interface';
@@ -21,6 +22,7 @@ export class IncomeComponent implements OnInit {
   @Input() view = [1000, 400];
   @Input() title = 'Net worth graph based on selected tabs';
   @Output() hidden: EventEmitter<any> = new EventEmitter;
+  @Output() loadPrevious: EventEmitter<any> = new EventEmitter;
 
   public isHidden = false;
   public visible = true;
@@ -110,8 +112,7 @@ export class IncomeComponent implements OnInit {
   }
 
   axisFormat(val) {
-    const stringDate: string = val.toTimeString().split(' ')[0];
-    return stringDate.substr(0, stringDate.length - 3);
+    return moment(val).format('LT');
   }
 
   setColorScheme(name) {
@@ -120,13 +121,18 @@ export class IncomeComponent implements OnInit {
   }
 
   select(data): void {
-    const items = this.dateData[0].series.filter(t => {
+    const snapshot = this.dateData[0].series.filter(t => {
       if (t.name === data.name) {
         return true;
       }
       return false;
-    })[0].items;
-    console.log('[INFO] Clicked item items: ', items);
+    })[0];
+    console.log('[INFO] Clicked snapshot: ', snapshot);
+    this.loadPreviousSnapshot(snapshot);
+  }
+
+  loadPreviousSnapshot(snapshot) {
+    this.loadPrevious.emit(snapshot);
   }
 
   onLegendLabelClick(entry) {