+++ ExilePartyClient/src/app/shared/providers/income.service.ts
@@ -1,10 +1,8 @@
-import 'rxjs/add/observable/empty';
 import 'rxjs/add/observable/forkJoin';
 import 'rxjs/add/observable/from';
 import 'rxjs/add/operator/concatMap';
 import 'rxjs/add/operator/delay';
 import 'rxjs/add/operator/do';
-import 'rxjs/add/operator/take';
 
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
@@ -15,7 +13,6 @@ import { Player } from '../interfaces/player.interface';
 import { NinjaLine, NinjaTypes } from '../interfaces/poe-ninja.interface';
 import { Stash } from '../interfaces/stash.interface';
 import { AccountService } from './account.service';
-import { AnalyticsService } from './analytics.service';
 import { ExternalService } from './external.service';
 import { LogService } from './log.service';
 import { NinjaService } from './ninja.service';
@@ -55,7 +52,7 @@ export class IncomeService {
 
     this.sessionId = sessionId;
 
-    this.netWorthHistory = this.settingsService.get('networth');
+    this.loadSnapshotsFromSettings();
 
     this.accountService.player.subscribe(res => {
       if (res !== undefined) {
@@ -63,6 +60,10 @@ export class IncomeService {
         this.localPlayer.netWorthSnapshots = this.netWorthHistory.history;
       }
     });
+  }
+
+  loadSnapshotsFromSettings() {
+    this.netWorthHistory = this.settingsService.get('networth');
   }
 
   Snapshot() {
@@ -81,6 +82,7 @@ export class IncomeService {
 
         this.netWorthHistory.history = this.netWorthHistory.history
           .filter((snaphot: NetWorthSnapshot) => snaphot.timestamp > oneHourAgo);
+
         // We are a new player that have not parsed income before
         // Remove the placeholder element
         if (
@@ -103,7 +105,6 @@ export class IncomeService {
         this.partyService.updatePlayer(this.localPlayer);
         this.logService.log('Finished Snapshotting player net worth');
 
-        // done snapshotting
         this.isSnapshotting = false;
       });
     }